Expand data to include years between inventory years

This is an "internal" function—most users will want to run fia_annualize() instead. This expands the data frame in preparation for interpolation of now "missing" values between inventory years or measurement years depending on the year variable selection. Time-invariant variables tree_ID, plot_ID, SPCD, MORTYR, ECOSUBCD, DESIGNCD, and PROP_BASIS are simply filled in with tidyr::fill(). Categorical variables STATUDSCD, RECONCILECD, STDORGCD, CONDID, and COND_STATUS_CD are modified to replace NAs with 999 so that they are properly interpolated by interpolate_data() (which converts them back to NAs).

Usage

expand_data(data, year_var = c("INVYR", "MEASYEAR"))

Arguments

data

tibble produced by fia_tidy()—must have at least tree_ID and INVYR columns.

year_var

character indicating which year variable to use; default is INVYR.

Value

a tibble with a logical column interpolated marking whether a row was present in the original data (FALSE) or was added (TRUE).