Assign interpolated data plots to estimation units and strata

This adds columns from various POP_* tables to annualized data so that it can be used for stratified estimation (see, for example, FIA Demystified). Each plot in each year is matched with an EVALID and it's associated data as such:

The EVALID must have both the "EXPVOL" and "EXPCURR" EVAL_TYP.

The year indicated by the middle two digits of the EVALID (usually, but not always the END_INVYR) must match the YEAR in the annualized data for that plot.

When there are gaps (e.g. because a plot was not sampled and not belong to an EVALID with "EXPVOL" or "EXPCURR") the EVALIDs are filled down, then up.

Usage

fia_assign_strata(data_annualized, db)

Arguments

data_annualized

Annualized data produced by fia_annualize().

db

The list of tables produced by fia_load().

Details

This means that the EVALID-associated data added by this function may be in conflict with the results of interpolation by fia_annualize()! When using this function to do stratified estimation, use the PLOT_STATUS_CD as part of the domain indicator to correctly exclude any non-sampled plots with no tree data!

See also

Examples

# Load example data included in package
db <- fia_load("RI", dir = system.file("exdata", package = "forestTIME"))

# Annualize data
data_annualized <- db |> fia_tidy() |>
   fia_annualize(use_mortyr = FALSE)
#> ℹ Adjusting for mortality
#> ℹ Interpolating between surveys
#> ℹ Expanding years between surveys
#> ℹ Wrangling data
#> ✔ Wrangling data [456ms]
#> 
#> ℹ Expanding years between surveys
#> ✔ Expanding years between surveys [4.6s]
#> 
#> ℹ Interpolating between surveys
#> ✔ Interpolating between surveys [16s]
#> 
#> ℹ Adjusting for mortality
#> ✔ Adjusting for mortality [18.5s]
#> 

# Assign plots to strata, estimation units, and EVLIDs
data_stratified <- fia_assign_strata(data_annualized, db)